# [Sub-issue 3/4] Monitoring: Comprehensive Performance Metrics Collection System