… src-tauri/src/vector_db/metrics_collector.rs
… src-tauri/src/vector_db/metrics_collector_tests.rs
… src-tauri/src/vector_db/monitored_search.rs